库存冻结解冻

+++ wms-entity/src/main/java/com/wms/entity/InvLotLocId.java
@@ -51,7 +51,7 @@ public class InvLotLocId  implements Serializable {
 
 	private Double price;
 
-	private Long onholdlocker;
+	private int onholdlocker;
 
 	private String lpn;
